test suite, recevoir ok envoyer bof

+++ API_cote_port_com/src/ecoute/translator.java
@@ -43,7 +43,7 @@ public class translator {
 		}
 		else
 		{
-			valeurs_envoie[OCTET_ODRE_ID] = new Byte(o.getIdDevice());
+			valeurs_envoie[OCTET_ODRE_ID] = (byte) new Byte(o.getIdDevice());
 			valeurs_envoie[OCTET_ORDRE_IDR] = 0;
 			valeurs_envoie[OCTET_ORDRE_NIVEAU] = (byte) ((o.getLevel_require()<<OFFSET_NIVEAU_ORDRE)|MASK_ORDRE);
 		}
@@ -53,11 +53,11 @@ public class translator {
 	public static Data bytesToData(byte[] b)
 	{
 			
-		int id = b[OCTET_ID];
-		int state = (b[OCTET_ETAT_POMPE]>>OFFSET_ETAT_POMPE) & MASK_ETAT_POMPE;
-		int levelMax = (b[OCTET_NOMBRE_DE_NIVEAU]>>OFFSET_NOMBRE_DE_NIVEAU) & MASK_NOMBRE_DE_NIVEAU;
-		int level = ((b[OCTET_NIVEAU_P1] & MASK_NIVEAU_P1)<<OFFSET_NIVEAU_P1)|((b[OCTET_NIVEAU_P2]>>OFFSET_NIVEAU_P2)& MASK_NIVEAU_P2);
-		int levelb = (b[OCTET_ETAT_BATERIE] & MASK_ETAT_BATERIE);
+		int id = (byte) b[OCTET_ID];
+		int state = (byte) (b[OCTET_ETAT_POMPE]>>OFFSET_ETAT_POMPE) & MASK_ETAT_POMPE;
+		int levelMax = (byte) (b[OCTET_NOMBRE_DE_NIVEAU]>>OFFSET_NOMBRE_DE_NIVEAU) & MASK_NOMBRE_DE_NIVEAU;
+		int level = (byte) ((b[OCTET_NIVEAU_P1] & MASK_NIVEAU_P1)<<OFFSET_NIVEAU_P1)|((b[OCTET_NIVEAU_P2]>>OFFSET_NIVEAU_P2)& MASK_NIVEAU_P2);
+		int levelb = (byte) (b[OCTET_ETAT_BATERIE] & MASK_ETAT_BATERIE);
 		
 		Data donnees = new Data(id,state,levelMax,level,levelb);
 		return donnees;
